refactor: create a small wrapper module for ohm
This moves ohm into a tiny wrapper module and prints an error message if
the git submodule hasn't been downloaded.
